move setup request logic to zkwasm_setup.js
