+++ api/requests/zkwasm_setup.js
@@ -1,0 +1,57 @@
+import FormData from "form-data";
+import axios from "axios";
+import url from "./url.js";
+import { Wallet } from "ethers";
+import { ZkWasmUtil } from "zkwasm-service-helper";
+import { config } from "../../config.js";
+
+export async function zkwasm_setup(
+    name,
+    image_md5,
+    image,
+    user_address,
+    description_url,
+    avator_url,
+    circuit_size
+    ){
+    let isSetUpSuccess = true;
+
+    let message = ZkWasmUtil.createAddImageSignMessage({
+      name: name,
+      image_md5: image_md5,
+      image: image,
+      user_address: user_address,
+      description_url: description_url,
+      avator_url: avator_url,
+      circuit_size: circuit_size,
+    });
+    const wallet = new Wallet(config.UserPrivateKey);
+    let signature = await wallet.signMessage(message);
+
+    let formData = new FormData();
+    formData.append("name", name);
+    formData.append("image_md5", image_md5);
+    formData.append("image", image);
+    formData.append("user_address", user_address);
+    formData.append("description_url", description_url);
+    formData.append("avator_url", avator_url);
+    formData.append("circuit_size", circuit_size);
+    formData.append("signature", signature);
+
+    let requestConfig = {
+      method: "post",
+      maxBodyLength: Infinity,
+      url: url.postNewWasmImage().url,
+      headers: {
+        ...formData.getHeaders(),
+      },
+      data: formData,
+    };
+
+    let errorMessage = "";
+    const response = await axios.request(requestConfig).catch((error) => {
+        isSetUpSuccess = false;
+        errorMessage = error.response.data;
+    });
+    return [response, isSetUpSuccess, errorMessage]
+}
